debugging addEntry

+++ src/trees/BSTNode.java
@@ -4,29 +4,29 @@ public class BSTNode {
     /*this node store int values
     * private to protected and set the modifiers through
     * defined methods*/
-    private int data;
+    private long data;
     public BSTNode leftNode, rightNode,  parent; //pointers to left, right and parent nodes
     //init the node with an int value and set the pointers to null
-    public BSTNode (int _num) {
+    public BSTNode (long _num) {
         this.data = _num;
         this.leftNode = null;
         this.rightNode = null;
         this.parent = null;
     }
     //get the data
-    public int getData() {
+    public long getData() {
         return data;
     }
     //get data from left node
-    public int getLeftNode() {
+    public long getLeftNode() {
         return leftNode.getData();
     }
     //get data from right node
-    public int getRightNode() {
+    public long getRightNode() {
         return rightNode.getData();
     }
     //setter to modify the value
-    public void setData(int _num) {
+    public void setData(long _num) {
         this.data = _num;
     }
     //setter to modify the left node value
